requirements.txt
python-dotenv

pymupdf

tiktoken

langchain
langchain-core
langchain-text-splitters
langchain-google-genai
langchain_qdrant

qdrant-client